requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssselect==1.2.0
cssutils==2.6.0
defusedxml==0.7.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
fast-captcha==0.1.3
fastapi==0.87.0
fastapi-redis-cache==0.2.5
fire==0.5.0
fonttools==4.39.4
fpdf2==2.7.4
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
husrev==1.5.3
idna==3.4
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
orjson==3.8.1
passlib==1.7.4
pdf2docx==0.5.6
Pillow==8.4.0
premailer==3.10.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
PyMuPDF==1.22.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
redis==4.5.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
termcolor==2.3.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
